Move validations to service

+++ src/main/java/com/mysari/radio/service/StationServiceImpl.java
@@ -3,12 +3,14 @@
  */
 package com.mysari.radio.service;
 
+import java.text.MessageFormat;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.mysari.radio.entity.Station;
+import com.mysari.radio.exception.StationException;
 import com.mysari.radio.mapper.StationMapper;
 
 /**
@@ -27,33 +29,63 @@ public class StationServiceImpl implements StationService {
 	}
 
 	@Override
-	public void addStation(Station station) {
+	public void addStation(Station station) throws StationException {
+		if (stationMapper.findByID(station.getStationId()) != null) {
+			String message = MessageFormat.format("Station for Id {0} already exists", station.getStationId());
+			throw new StationException(message);
+		}
 		stationMapper.addStation(station);
 	}
 
 	@Override
-	public void removeStation(String stationId) {
+	public void removeStation(String stationId) throws StationException {
+		if (null == stationMapper.findByID(stationId)) {
+			String message = MessageFormat.format("Station for Id {0} does not exist", stationId);
+			throw new StationException(message);
+		}
 		stationMapper.removeStation(stationId);
 	}
 
 	@Override
-	public void updateStation(Station station) {
+	public void updateStation(Station station, String stationId) throws StationException {
+		if (null != station.getStationId() && !station.getStationId().equalsIgnoreCase(stationId)) {
+			throw new StationException("StationId Mismatch");
+		}
+		if (null == stationMapper.findByID(stationId)) {
+			String message = MessageFormat.format("Station for Id {0} does not exist", stationId);
+			throw new StationException(message);
+		}
 		stationMapper.updateStation(station.getStationId(), station.getName(), station.isHd_Enabled(),
 				station.getCallSign());
 	}
 
 	@Override
-	public Station findByID(String stationId) {
-		return stationMapper.findByID(stationId);
+	public Station findByID(String stationId) throws StationException {
+		final Station station = stationMapper.findByID(stationId);
+		if (null == station) {
+			String message = MessageFormat.format("Station for Id {0} does not exist", stationId);
+			throw new StationException(message);
+		}
+		return station;
 	}
 
 	@Override
-	public Station findByName(String name) {
-		return stationMapper.findByName(name);
+	public Station findByName(String name) throws StationException {
+		final Station station = stationMapper.findByName(name);
+		if (null == station) {
+			String message = MessageFormat.format("Station for Name {0} does not exist", name);
+			throw new StationException(message);
+		}
+		return station;
 	}
 
 	@Override
-	public List<Station> findByIdorName(String search) {
+	public List<Station> findByIdorName(String search) throws StationException {
+		final List<Station> station = stationMapper.findByIDorName(search);
+		if (null == station || station.isEmpty()) {
+			String message = MessageFormat.format("Station for search {0} does not exist", search);
+			throw new StationException(message);
+		}
 		return stationMapper.findByIDorName(search);
 	}
